+++ Source/CesiumRuntime/Private/CesiumGltfComponent.h
@@ -36,10 +36,13 @@ struct FRasterOverlayTile {
   GENERATED_BODY()
 
   UPROPERTY()
+  FString OverlayName;
+
+  UPROPERTY()
   UTexture2D* Texture;
 
-  FLinearColor TextureCoordinateRectangle;
   FLinearColor TranslationAndScale;
+  int32 TextureCoordinateID;
 };
 
 UCLASS()
@@ -62,8 +65,7 @@ public:
       std::unique_ptr<HalfConstructed> HalfConstructed,
       const glm::dmat4x4& CesiumToUnrealTransform,
       UMaterialInterface* BaseMaterial,
-      UMaterialInterface* BaseWaterMaterial,
-      UMaterialInterface* BaseOpacityMaterial);
+      UMaterialInterface* BaseWaterMaterial);
 
   UCesiumGltfComponent();
   virtual ~UCesiumGltfComponent();
@@ -74,24 +76,20 @@ public:
   UPROPERTY(EditAnywhere, Category = "Cesium")
   UMaterialInterface* BaseMaterialWithWater;
 
-  UPROPERTY(EditAnywhere, Category = "Cesium")
-  UMaterialInterface* OpacityMaskMaterial;
-
   void UpdateTransformFromCesium(const glm::dmat4& CesiumToUnrealTransform);
 
   void AttachRasterTile(
       const Cesium3DTilesSelection::Tile& Tile,
       const Cesium3DTilesSelection::RasterOverlayTile& RasterTile,
       UTexture2D* Texture,
-      const CesiumGeometry::Rectangle& TextureCoordinateRectangle,
       const glm::dvec2& Translation,
-      const glm::dvec2& Scale);
+      const glm::dvec2& Scale,
+      int32_t TextureCoordinateID);
 
   void DetachRasterTile(
       const Cesium3DTilesSelection::Tile& Tile,
       const Cesium3DTilesSelection::RasterOverlayTile& RasterTile,
-      UTexture2D* Texture,
-      const CesiumGeometry::Rectangle& TextureCoordinateRectangle);
+      UTexture2D* Texture);
 
   UFUNCTION(BlueprintCallable, Category = "Collision")
   virtual void SetCollisionEnabled(ECollisionEnabled::Type NewType);
@@ -99,7 +97,6 @@ public:
   virtual void FinishDestroy() override;
 
 private:
-  void UpdateRasterOverlays();
-
-  TArray<FRasterOverlayTile> OverlayTiles;
+  UPROPERTY()
+  UTexture2D* Transparent1x1;
 };