+++ Source/CesiumRuntime/Public/Cesium3DTileset.h
@@ -19,6 +19,7 @@
 #include "Cesium3DTileset.generated.h"
 
 class UMaterialInterface;
+class ACesiumCartographicSelection;
 
 namespace Cesium3DTilesSelection {
 class Tileset;
@@ -243,6 +244,15 @@ public:
   float CulledScreenSpaceError = 64.0;
 
   /**
+   * Refreshes this tileset, ensuring that all materials and other settings are
+   * applied. It is not usually necessary to invoke this, but when
+   * behind-the-scenes changes are made and not reflected in the tileset, this
+   * function can help.
+   */
+  UFUNCTION(CallInEditor, BlueprintCallable, Category = "Cesium")
+  void RefreshTileset();
+
+  /**
    * Pauses level-of-detail and culling updates of this tileset.
    */
   UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Cesium|Debug")
@@ -415,20 +425,6 @@ private:
       Category = "Cesium|Rendering")
   UMaterialInterface* WaterMaterial = nullptr;
 
-  /**
-   * A custom Material to use to render this tileset in areas using opacity
-   * masks, in order to implement custom visual effects.
-   *
-   * The custom material should generally be created by copying the
-   * "M_CesiumDefaultMasked" material and customizing it as desired.
-   */
-  UPROPERTY(
-      EditAnywhere,
-      BlueprintGetter = GetOpacityMaskMaterial,
-      BlueprintSetter = SetOpacityMaskMaterial,
-      Category = "Cesium|Rendering")
-  UMaterialInterface* OpacityMaskMaterial = nullptr;
-
 protected:
   UPROPERTY()
   FString PlatformName;
@@ -493,14 +489,6 @@ public:
 
   UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
   void SetWaterMaterial(UMaterialInterface* InMaterial);
-
-  UFUNCTION(BlueprintGetter, Category = "Cesium|Rendering")
-  UMaterialInterface* GetOpacityMaskMaterial() const {
-    return OpacityMaskMaterial;
-  }
-
-  UFUNCTION(BlueprintSetter, Category = "Cesium|Rendering")
-  void SetOpacityMaskMaterial(UMaterialInterface* InMaterial);
 
   UFUNCTION(BlueprintCallable, Category = "Cesium|Rendering")
   void PlayMovieSequencer();