+++ Source/CesiumRuntime/Public/CesiumRasterOverlay.h
@@ -22,6 +22,18 @@ class CESIUMRUNTIME_API UCesiumRasterOverlay : public UActorComponent {
   GENERATED_BODY()
 
 public:
+  /**
+   * The key to use to match this overlay to a material layer.
+   *
+   * When using Material Layers, any material layers inside a "Cesium" layer
+   * stack with a name that matches this name will have their Texture,
+   * TranslationScale, and TextureCoordinateIndex properties set automatically
+   * so that a ML_CesiumOverlay layer function (or similar) will correctly
+   * sample from this overlay.
+   */
+  UPROPERTY(EditAnywhere, Category = "Cesium")
+  FString MaterialLayerKey = "Overlay0";
+
   // Sets default values for this component's properties
   UCesiumRasterOverlay();